requirements.txt
# Core Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
pymongo==4.6.1
sqlalchemy==2.0.25

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Data Ingestion
requests==2.31.0
feedparser==6.0.11
python-dateutil==2.8.2

# AI/ML - Phase 2
transformers==4.36.0
torch==2.1.0
sentencepiece==0.1.99
accelerate==0.25.0
scikit-learn==1.4.0

# Phase 2: AI Explanations (Optional)
google-generativeai==0.3.2  # Gemini API for natural language explanations

# Phase 3: Dashboard (Streamlit)
streamlit==1.29.0
plotly==5.18.0
folium==0.15.1
streamlit-folium==0.16.0

# Web Scraping - Phase 2
beautifulsoup4==4.12.0
selenium==4.16.0
lxml==5.1.0
webdriver-manager==4.0.1

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3

# Development
pytest==7.4.4
black==24.1.1
